+++ src/frontend/js/widgets/CourseProductItem/components/SaleTunnel/index.tsx
@@ -4,6 +4,9 @@ import { Modal } from 'components/Modal';
 import type * as Joanie from 'types/Joanie';
 import { useOrders } from 'hooks/useOrders';
 import { IconTypeEnum } from 'components/Icon';
+import WebAnalyticsAPIHandler from 'api/web-analytics';
+import { CourseProductEvent } from 'types/web-analytics';
+import { useCourseProduct } from 'widgets/CourseProductItem/contexts/CourseProductContext';
 import { StepBreadcrumb } from '../StepBreadcrumb';
 import { Manifest, useStepManager } from '../../hooks/useStepManager';
 import { SaleTunnelStepResume } from '../SaleTunnelStepResume';
@@ -46,6 +49,7 @@ type Props = {
 const SaleTunnel = ({ product, isOpen = false, onClose }: Props) => {
   const intl = useIntl();
   const { methods: ordersMethods } = useOrders();
+  const { key } = useCourseProduct();
 
   const manifest: Manifest<TunnelSteps, 'resume'> = {
     start: 'validation',
@@ -59,11 +63,20 @@ const SaleTunnel = ({ product, isOpen = false, onClose }: Props) => {
         icon: IconTypeEnum.CREDIT_CARD,
         label: intl.formatMessage(messages.stepPayment),
         next: 'resume',
+        onEnter: () => {
+          WebAnalyticsAPIHandler()?.sendCourseProductEvent(
+            CourseProductEvent.PAYMENT_STEP_DISPLAYED,
+            key,
+          );
+        },
       },
       resume: {
         icon: IconTypeEnum.CHECK,
         label: intl.formatMessage(messages.stepResume),
         next: null,
+        onEnter: () => {
+          WebAnalyticsAPIHandler()?.sendCourseProductEvent(CourseProductEvent.PAYMENT_SUCCEED, key);
+        },
         onExit: () => {
           // Once the user has completed the purchase, we need to refetch the orders
           // to update the ordersQuery cache
@@ -77,9 +90,12 @@ const SaleTunnel = ({ product, isOpen = false, onClose }: Props) => {
 
   const modalRef = useRef<HTMLDivElement | null>(null);
 
-  const handleModalClose = () => {
-    reset();
+  const handleModalClose = (track: boolean = false) => {
+    if (track) {
+      WebAnalyticsAPIHandler()?.sendCourseProductEvent(CourseProductEvent.CLOSE_SALE_TUNNEL, key);
+    }
     onClose();
+    reset();
   };
 
   /**
@@ -100,13 +116,14 @@ const SaleTunnel = ({ product, isOpen = false, onClose }: Props) => {
       className="SaleTunnel__modal"
       isOpen={isOpen}
       onAfterOpen={(options) => {
+        WebAnalyticsAPIHandler()?.sendCourseProductEvent(CourseProductEvent.OPEN_SALE_TUNNEL, key);
         if (!options) {
           return;
         }
         focusCurrentStep(options.contentEl);
       }}
       contentRef={(ref) => (modalRef.current = ref)}
-      onRequestClose={handleModalClose}
+      onRequestClose={() => handleModalClose(true)}
       shouldCloseOnOverlayClick={false}
       shouldCloseOnEsc={false}
       testId="SaleTunnel__modal"